+++ example/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "editor.tabCompletion": "on",
+    "diffEditor.codeLens": true
+}
