+++ example/lib/loading.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class LoadingPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: CircularProgressIndicator(),
+      ),
+    );
+  }
+}
